Stop casting mmc_token return values to char
When mmc_token returns a token value it ors it with MMC_DATA_TOKEN_RESP
which only effects the most significant byte of the word. To ignore this
the result was cast to a char before comparison now MMC_DATA_TOKEN_RESP is
ored with the comparison value. This makes sure that the returned value is
actualy a data token response and not something else.